+++ PKHeX.Core/Legality/Tables/Tables6.cs
@@ -150,7 +150,7 @@ public static partial class Legal
 
     #endregion
 
-    internal static readonly HashSet<int> ValidMet_XY = new()
+    internal static readonly HashSet<ushort> ValidMet_XY = new()
     {
         006, 008, 009, 010, 012, 013, 014, 016, 017, 018, 020, 021, 022, 024, 026, 028, 029, 030, 032, 034, 035, 036,
         038, 039, 040, 042, 043, 044, 046, 047, 048, 050, 051, 052, 054, 055, 056, 058, 060, 062, 063, 064, 066, 067,
@@ -159,7 +159,7 @@ public static partial class Legal
         138, 140, 142, 144, 146, 148, 150, 152, 154, 156, 158, 160, 162, 164, 166, 168,
     };
 
-    internal static readonly HashSet<int> ValidMet_AO = new()
+    internal static readonly HashSet<ushort> ValidMet_AO = new()
     {
         170, 172, 174, 176, 178, 180, 182, 184, 186, 188, 190, 192, 194, 196, 198,
         200, 202, 204, 206, 208, 210, 212, 214, 216, 218, 220, 222, 224, 226, 228, 230, 232, 234, 236, 238, 240, 242,
